+++ CHANGELOG.md
@@ -28,6 +28,11 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Fixed
 
+- `kettle-family release --execute` runs release members sequentially on
+  TruffleRuby to avoid a TruffleRuby 24.2 internal `ENV.replace` crash from
+  `Bundler.with_unbundled_env` inside parallel release threads
+  ([truffleruby/truffleruby#4352](https://github.com/truffleruby/truffleruby/issues/4352)).
+
 ### Security
 
 ## [0.1.32] - 2026-07-01